Struct rusoto_ecs::CreateServiceRequest

pub struct CreateServiceRequest {
    pub client_token: Option<String>,
    pub cluster: Option<String>,
    pub deployment_configuration: Option<DeploymentConfiguration>,
    pub desired_count: i64,
    pub load_balancers: Option<Vec<LoadBalancer>>,
    pub placement_constraints: Option<Vec<PlacementConstraint>>,
    pub placement_strategy: Option<Vec<PlacementStrategy>>,
    pub role: Option<String>,
    pub service_name: String,
    pub task_definition: String,
}

Fields
client_token: Option<String>

Unique, case-sensitive identifier you provide to ensure the idempotency of the request. Up to 32 ASCII characters are allowed.

cluster: Option<String>

The short name or full Amazon Resource Name (ARN) of the cluster on which to run your service. If you do not specify a cluster, the default cluster is assumed.

deployment_configuration: Option<DeploymentConfiguration>

Optional deployment parameters that control how many tasks run during the deployment and the ordering of stopping and starting tasks.

desired_count: i64

The number of instantiations of the specified task definition to place and keep running on your cluster.

load_balancers: Option<Vec<LoadBalancer>>

A load balancer object representing the load balancer to use with your service. Currently, you are limited to one load balancer or target group per service. After you create a service, the load balancer name or target group ARN, container name, and container port specified in the service definition are immutable.

For Elastic Load Balancing Classic load balancers, this object must contain the load balancer name, the container name (as it appears in a container definition), and the container port to access from the load balancer. When a task from this service is placed on a container instance, the container instance is registered with the load balancer specified here.

For Elastic Load Balancing Application load balancers, this object must contain the load balancer target group ARN, the container name (as it appears in a container definition), and the container port to access from the load balancer. When a task from this service is placed on a container instance, the container instance and port combination is registered as a target in the target group specified here.

placement_constraints: Option<Vec<PlacementConstraint>>

An array of placement constraint objects to use for tasks in your service. You can specify a maximum of 10 constraints per task (this limit includes constraints in the task definition and those specified at run time).

placement_strategy: Option<Vec<PlacementStrategy>>

The placement strategy objects to use for tasks in your service. You can specify a maximum of 5 strategy rules per service.

role: Option<String>

The name or full Amazon Resource Name (ARN) of the IAM role that allows Amazon ECS to make calls to your load balancer on your behalf. This parameter is required if you are using a load balancer with your service. If you specify the role parameter, you must also specify a load balancer object with the loadBalancers parameter.

If your specified role has a path other than /, then you must either specify the full role ARN (this is recommended) or prefix the role name with the path. For example, if a role with the name bar has a path of /foo/ then you would specify /foo/bar as the role name. For more information, see Friendly Names and Paths in the IAM User Guide.

service_name: String

The name of your service. Up to 255 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed. Service names must be unique within a cluster, but you can have similarly named services in multiple clusters within a region or across multiple regions.

task_definition: String

The family and revision (family:revision) or full Amazon Resource Name (ARN) of the task definition to run in your service. If a revision is not specified, the latest ACTIVE revision is used.