Struct rusoto_ecs::EcsClient

pub struct EcsClient { /* fields omitted */ }

A client for the Amazon ECS API.

Methods

impl EcsClient

pub fn new(region: Region) -> EcsClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> EcsClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,