Struct rusoto_ecs::PlacementConstraint

pub struct PlacementConstraint {
    pub expression: Option<String>,
    pub type_: Option<String>,
}

An object representing a constraint on task placement. For more information, see Task Placement Constraints in the Amazon EC2 Container Service Developer Guide.

Fields
expression: Option<String>

A cluster query language expression to apply to the constraint. Note you cannot specify an expression if the constraint type is distinctInstance. For more information, see Cluster Query Language in the Amazon EC2 Container Service Developer Guide.

type_: Option<String>

The type of constraint. Use distinctInstance to ensure that each task in a particular group is running on a different container instance. Use memberOf to restrict selection to a group of valid candidates. Note that distinctInstance is not supported in task definitions.