Struct rusoto_ecs::Task

pub struct Task {
    pub attachments: Option<Vec<Attachment>>,
    pub cluster_arn: Option<String>,
    pub connectivity: Option<String>,
    pub connectivity_at: Option<f64>,
    pub container_instance_arn: Option<String>,
    pub containers: Option<Vec<Container>>,
    pub cpu: Option<String>,
    pub created_at: Option<f64>,
    pub desired_status: Option<String>,
    pub execution_stopped_at: Option<f64>,
    pub group: Option<String>,
    pub last_status: Option<String>,
    pub launch_type: Option<String>,
    pub memory: Option<String>,
    pub overrides: Option<TaskOverride>,
    pub platform_version: Option<String>,
    pub pull_started_at: Option<f64>,
    pub pull_stopped_at: Option<f64>,
    pub started_at: Option<f64>,
    pub started_by: Option<String>,
    pub stopped_at: Option<f64>,
    pub stopped_reason: Option<String>,
    pub stopping_at: Option<f64>,
    pub task_arn: Option<String>,
    pub task_definition_arn: Option<String>,
    pub version: Option<i64>,
}

Details on a task in a cluster.

Fields
attachments: Option<Vec<Attachment>>

The Elastic Network Adapter associated with the task if the task uses the awsvpc network mode.

cluster_arn: Option<String>

The ARN of the cluster that hosts the task.

connectivity: Option<String>

The connectivity status of a task.

connectivity_at: Option<f64>

The Unix time stamp for when the task last went into CONNECTED status.

container_instance_arn: Option<String>

The ARN of the container instances that host the task.

containers: Option<Vec<Container>>

The containers associated with the task.

cpu: Option<String>

The number of cpu units used by the task. If using the EC2 launch type, this field is optional and any value can be used. If using the Fargate launch type, this field is required and you must use one of the following values, which determines your range of valid values for the memory parameter:

256 (.25 vCPU) - Available memory values: 0.5GB, 1GB, 2GB

512 (.5 vCPU) - Available memory values: 1GB, 2GB, 3GB, 4GB

1024 (1 vCPU) - Available memory values: 2GB, 3GB, 4GB, 5GB, 6GB, 7GB, 8GB

2048 (2 vCPU) - Available memory values: Between 4GB and 16GB in 1GB increments

4096 (4 vCPU) - Available memory values: Between 8GB and 30GB in 1GB increments

created_at: Option<f64>

The Unix time stamp for when the task was created (the task entered the PENDING state).

desired_status: Option<String>

The desired status of the task.

execution_stopped_at: Option<f64>

The Unix timestamp for when the task execution stopped.

group: Option<String>

The name of the task group associated with the task.

last_status: Option<String>

The last known status of the task.

launch_type: Option<String>

The launch type on which your task is running.

memory: Option<String>

The amount (in MiB) of memory used by the task. If using the EC2 launch type, this field is optional and any value can be used. If using the Fargate launch type, this field is required and you must use one of the following values, which determines your range of valid values for the cpu parameter:

0.5GB, 1GB, 2GB - Available cpu values: 256 (.25 vCPU)

1GB, 2GB, 3GB, 4GB - Available cpu values: 512 (.5 vCPU)

2GB, 3GB, 4GB, 5GB, 6GB, 7GB, 8GB - Available cpu values: 1024 (1 vCPU)

Between 4GB and 16GB in 1GB increments - Available cpu values: 2048 (2 vCPU)

Between 8GB and 30GB in 1GB increments - Available cpu values: 4096 (4 vCPU)

overrides: Option<TaskOverride>

One or more container overrides.

platform_version: Option<String>

The platform version on which your task is running. For more information, see AWS Fargate Platform Versions in the Amazon Elastic Container Service Developer Guide.

pull_started_at: Option<f64>

The Unix time stamp for when the container image pull began.

pull_stopped_at: Option<f64>

The Unix time stamp for when the container image pull completed.

started_at: Option<f64>

The Unix time stamp for when the task started (the task transitioned from the PENDING state to the RUNNING state).

started_by: Option<String>

The tag specified when a task is started. If the task is started by an Amazon ECS service, then the startedBy parameter contains the deployment ID of the service that starts it.

stopped_at: Option<f64>

The Unix time stamp for when the task was stopped (the task transitioned from the RUNNING state to the STOPPED state).

stopped_reason: Option<String>

The reason the task was stopped.

stopping_at: Option<f64>

The Unix time stamp for when the task will stop (the task transitioned from the RUNNING state to the STOPPED state).

task_arn: Option<String>

The Amazon Resource Name (ARN) of the task.

task_definition_arn: Option<String>

The ARN of the task definition that creates the task.

version: Option<i64>

The version counter for the task. Every time a task experiences a change that triggers a CloudWatch event, the version counter is incremented. If you are replicating your Amazon ECS task state with CloudWatch Events, you can compare the version of a task reported by the Amazon ECS APIs with the version reported in CloudWatch Events for the task (inside the detail object) to verify that the version in your event stream is current.