Struct rusoto_ecs::TaskDefinition

pub struct TaskDefinition {
    pub container_definitions: Option<Vec<ContainerDefinition>>,
    pub family: Option<String>,
    pub network_mode: Option<String>,
    pub placement_constraints: Option<Vec<TaskDefinitionPlacementConstraint>>,
    pub requires_attributes: Option<Vec<Attribute>>,
    pub revision: Option<i64>,
    pub status: Option<String>,
    pub task_definition_arn: Option<String>,
    pub task_role_arn: Option<String>,
    pub volumes: Option<Vec<Volume>>,
}

Details of a task definition.

Fields
container_definitions: Option<Vec<ContainerDefinition>>

A list of container definitions in JSON format that describe the different containers that make up your task. For more information about container definition parameters and defaults, see Amazon ECS Task Definitions in the Amazon EC2 Container Service Developer Guide.

family: Option<String>

The family of your task definition, used as the definition name.

network_mode: Option<String>

The Docker networking mode to use for the containers in the task. The valid values are none, bridge, and host.

If the network mode is none, the containers do not have external connectivity. The default Docker network mode is bridge. The host network mode offers the highest networking performance for containers because it uses the host network stack instead of the virtualized network stack provided by the bridge mode.

For more information, see Network settings in the Docker run reference.

placement_constraints: Option<Vec<TaskDefinitionPlacementConstraint>>

An array of placement constraint objects to use for tasks.

requires_attributes: Option<Vec<Attribute>>

The container instance attributes required by your task.

revision: Option<i64>

The revision of the task in a particular family. The revision is a version number of a task definition in a family. When you register a task definition for the first time, the revision is 1; each time you register a new revision of a task definition in the same family, the revision value always increases by one (even if you have deregistered previous revisions in this family).

status: Option<String>

The status of the task definition.

task_definition_arn: Option<String>

The full Amazon Resource Name (ARN) of the task definition.

task_role_arn: Option<String>

The Amazon Resource Name (ARN) of the IAM role that containers in this task can assume. All containers in this task are granted the permissions that are specified in this role.

volumes: Option<Vec<Volume>>

The list of volumes in a task. For more information about volume definition parameters and defaults, see Amazon ECS Task Definitions in the Amazon EC2 Container Service Developer Guide.