Struct rusoto_ecs::TaskSet

pub struct TaskSet {
    pub cluster_arn: Option<String>,
    pub computed_desired_count: Option<i64>,
    pub created_at: Option<f64>,
    pub external_id: Option<String>,
    pub id: Option<String>,
    pub launch_type: Option<String>,
    pub load_balancers: Option<Vec<LoadBalancer>>,
    pub network_configuration: Option<NetworkConfiguration>,
    pub pending_count: Option<i64>,
    pub platform_version: Option<String>,
    pub running_count: Option<i64>,
    pub scale: Option<Scale>,
    pub service_arn: Option<String>,
    pub service_registries: Option<Vec<ServiceRegistry>>,
    pub stability_status: Option<String>,
    pub stability_status_at: Option<f64>,
    pub started_by: Option<String>,
    pub status: Option<String>,
    pub task_definition: Option<String>,
    pub task_set_arn: Option<String>,
    pub updated_at: Option<f64>,
}

Information about a set of Amazon ECS tasks in either an AWS CodeDeploy or an EXTERNAL deployment. An Amazon ECS task set includes details such as the desired number of tasks, how many tasks are running, and whether the task set serves production traffic.

Fields
cluster_arn: Option<String>

The Amazon Resource Name (ARN) of the cluster that the service that hosts the task set exists in.

computed_desired_count: Option<i64>

The computed desired count for the task set. This is calculated by multiplying the service's desiredCount by the task set's scale percentage. The result is always rounded up. For example, if the computed desired count is 1.2, it rounds up to 2 tasks.

created_at: Option<f64>

The Unix timestamp for when the task set was created.

external_id: Option<String>

The external ID associated with the task set.

If a task set is created by an AWS CodeDeploy deployment, the externalId parameter contains the AWS CodeDeploy deployment ID.

If a task set is created for an external deployment and is associated with a service discovery registry, the externalId parameter contains the ECS_TASK_SET_EXTERNAL_ID AWS Cloud Map attribute.

id: Option<String>

The ID of the task set.

launch_type: Option<String>

The launch type the tasks in the task set are using. For more information, see Amazon ECS Launch Types in the Amazon Elastic Container Service Developer Guide.

load_balancers: Option<Vec<LoadBalancer>>

Details on a load balancer that is used with a task set.

network_configuration: Option<NetworkConfiguration>

The network configuration for the task set.

pending_count: Option<i64>

The number of tasks in the task set that are in the PENDING status during a deployment. A task in the PENDING state is preparing to enter the RUNNING state. A task set enters the PENDING status when it launches for the first time or when it is restarted after being in the STOPPED state.

platform_version: Option<String>

The platform version on which the tasks in the task set are running. A platform version is only specified for tasks using the Fargate launch type. If one is not specified, the LATEST platform version is used by default. For more information, see AWS Fargate Platform Versions in the Amazon Elastic Container Service Developer Guide.

running_count: Option<i64>

The number of tasks in the task set that are in the RUNNING status during a deployment. A task in the RUNNING state is running and ready for use.

scale: Option<Scale>

A floating-point percentage of the desired number of tasks to place and keep running in the task set.

service_arn: Option<String>

The Amazon Resource Name (ARN) of the service the task set exists in.

service_registries: Option<Vec<ServiceRegistry>>

The details of the service discovery registries to assign to this task set. For more information, see Service Discovery.

stability_status: Option<String>

The stability status, which indicates whether the task set has reached a steady state. If the following conditions are met, the task set will be in STEADY_STATE:

The task runningCount is equal to the computedDesiredCount.

The pendingCount is 0.

There are no tasks running on container instances in the DRAINING status.

All tasks are reporting a healthy status from the load balancers, service discovery, and container health checks.

If a healthCheckGracePeriodSeconds value was set when the service was created, you may see a STEADY_STATE reached since unhealthy Elastic Load Balancing target health checks will be ignored until it expires.

If any of those conditions are not met, the stability status returns STABILIZING.

stability_status_at: Option<f64>

The Unix timestamp for when the task set stability status was retrieved.

started_by: Option<String>

The tag specified when a task set is started. If the task set is created by an AWS CodeDeploy deployment, the startedBy parameter is CODE_DEPLOY. For a task set created for an external deployment, the startedBy field isn't used.

status: Option<String>

The status of the task set. The following describes each state:

PRIMARY

The task set is serving production traffic.

ACTIVE

The task set is not serving production traffic.

DRAINING

The tasks in the task set are being stopped and their corresponding targets are being deregistered from their target group.

task_definition: Option<String>

The task definition the task set is using.

task_set_arn: Option<String>

The Amazon Resource Name (ARN) of the task set.

updated_at: Option<f64>

The Unix timestamp for when the task set was last updated.