Struct rusoto_ecs::UpdateServiceRequest

pub struct UpdateServiceRequest {
    pub cluster: Option<String>,
    pub deployment_configuration: Option<DeploymentConfiguration>,
    pub desired_count: Option<i64>,
    pub force_new_deployment: Option<bool>,
    pub health_check_grace_period_seconds: Option<i64>,
    pub network_configuration: Option<NetworkConfiguration>,
    pub platform_version: Option<String>,
    pub service: String,
    pub task_definition: Option<String>,
}

Fields
cluster: Option<String>

The short name or full Amazon Resource Name (ARN) of the cluster that your service is running on. If you do not specify a cluster, the default cluster is assumed.

deployment_configuration: Option<DeploymentConfiguration>

Optional deployment parameters that control how many tasks run during the deployment and the ordering of stopping and starting tasks.

desired_count: Option<i64>

The number of instantiations of the task to place and keep running in your service.

force_new_deployment: Option<bool>

Whether or not to force a new deployment of the service.

health_check_grace_period_seconds: Option<i64>

The period of time, in seconds, that the Amazon ECS service scheduler should ignore unhealthy Elastic Load Balancing target health checks after a task has first started. This is only valid if your service is configured to use a load balancer. If your service's tasks take a while to start and respond to ELB health checks, you can specify a health check grace period of up to 1,800 seconds during which the ECS service scheduler will ignore ELB health check status. This grace period can prevent the ECS service scheduler from marking tasks as unhealthy and stopping them before they have time to come up.

network_configuration: Option<NetworkConfiguration>

The network configuration for the service. This parameter is required for task definitions that use the awsvpc network mode to receive their own Elastic Network Interface, and it is not supported for other network modes. For more information, see Task Networking in the Amazon Elastic Container Service Developer Guide.

Updating a service to add a subnet to a list of existing subnets does not trigger a service deployment. For example, if your network configuration change is to keep the existing subnets and simply add another subnet to the network configuration, this does not trigger a new service deployment.

platform_version: Option<String>

The platform version you want to update your service to run.

service: String

The name of the service to update.

task_definition: Option<String>

The family and revision (family:revision) or full ARN of the task definition to run in your service. If a revision is not specified, the latest ACTIVE revision is used. If you modify the task definition with UpdateService, Amazon ECS spawns a task with the new version of the task definition and then stops an old task after the new version is running.