Crate rusoto_ecs

Amazon Elastic Container Service (Amazon ECS) is a highly scalable, fast, container management service that makes it easy to run, stop, and manage Docker containers on a cluster. You can host your cluster on a serverless infrastructure that is managed by Amazon ECS by launching your services or tasks using the Fargate launch type. For more control, you can host your tasks on a cluster of Amazon Elastic Compute Cloud (Amazon EC2) instances that you manage by using the EC2 launch type. For more information about launch types, see Amazon ECS Launch Types.

Amazon ECS lets you launch and stop container-based applications with simple API calls, allows you to get the state of your cluster from a centralized service, and gives you access to many familiar Amazon EC2 features.

You can use Amazon ECS to schedule the placement of containers across your cluster based on your resource needs, isolation policies, and availability requirements. Amazon ECS eliminates the need for you to operate your own cluster management and configuration management systems or worry about scaling your management infrastructure.

If you're using the service, you're probably looking for EcsClient and Ecs.

Structs

Attachment
An object representing a container instance or task attachment.

AttachmentStateChange
An object representing a change in state for a task attachment.

Attribute
An attribute is a name-value pair associated with an Amazon ECS object. Attributes enable you to extend the Amazon ECS data model by adding custom metadata to your resources. For more information, see Attributes in the Amazon Elastic Container Service Developer Guide.

AwsVpcConfiguration
An object representing the networking details for a task or service.

Cluster
A regional grouping of one or more container instances on which you can run task requests. Each account receives a default cluster the first time you use the Amazon ECS service, but you may also create other clusters. Clusters may contain more than one instance type simultaneously.

Container
A Docker container that is part of a task.

ContainerDefinition
Container definitions are used in task definitions to describe the different containers that are launched as part of a task.

ContainerInstance
An EC2 instance that is running the Amazon ECS agent and has been registered with a cluster.

ContainerOverride
The overrides that should be sent to a container.

ContainerStateChange
An object representing a change in state for a container.

CreateClusterRequest

CreateClusterResponse

CreateServiceRequest

CreateServiceResponse

DeleteAttributesRequest

DeleteAttributesResponse

DeleteClusterRequest

DeleteClusterResponse

DeleteServiceRequest

DeleteServiceResponse

Deployment
The details of an Amazon ECS service deployment.

DeploymentConfiguration
Optional deployment parameters that control how many tasks run during the deployment and the ordering of stopping and starting tasks.

DeregisterContainerInstanceRequest

DeregisterContainerInstanceResponse

DeregisterTaskDefinitionRequest

DeregisterTaskDefinitionResponse

DescribeClustersRequest

DescribeClustersResponse

DescribeContainerInstancesRequest

DescribeContainerInstancesResponse

DescribeServicesRequest

DescribeServicesResponse

DescribeTaskDefinitionRequest

DescribeTaskDefinitionResponse

DescribeTasksRequest

DescribeTasksResponse

Device
An object representing a container instance host device.

DiscoverPollEndpointRequest

DiscoverPollEndpointResponse

EcsClient
A client for the Amazon ECS API.

Failure
A failed resource.

HealthCheck
An object representing a container health check. Health check parameters that are specified in a container definition override any Docker health checks that exist in the container image (such as those specified in a parent image or from the image's Dockerfile).

HostEntry
Hostnames and IP address entries that are added to the /etc/hosts file of a container via the extraHosts parameter of its ContainerDefinition.

HostVolumeProperties
Details on a container instance host volume.

KernelCapabilities
The Linux capabilities for the container that are added to or dropped from the default configuration provided by Docker. For more information on the default capabilities and the non-default available capabilities, see Runtime privilege and Linux capabilities in the Docker run reference. For more detailed information on these Linux capabilities, see the capabilities(7) Linux manual page.

KeyValuePair
A key and value pair object.

LinuxParameters
Linux-specific options that are applied to the container, such as Linux KernelCapabilities.

ListAttributesRequest

ListAttributesResponse

ListClustersRequest

ListClustersResponse

ListContainerInstancesRequest

ListContainerInstancesResponse

ListServicesRequest

ListServicesResponse

ListTaskDefinitionFamiliesRequest

ListTaskDefinitionFamiliesResponse

ListTaskDefinitionsRequest

ListTaskDefinitionsResponse

ListTasksRequest

ListTasksResponse

LoadBalancer
Details on a load balancer that is used with a service.

Services with tasks that use the awsvpc network mode (for example, those with the Fargate launch type) only support Application Load Balancers and Network Load Balancers; Classic Load Balancers are not supported. Also, when you create any target groups for these services, you must choose ip as the target type, not instance, because tasks that use the awsvpc network mode are associated with an elastic network interface, not an Amazon EC2 instance.

LogConfiguration
Log configuration options to send to a custom log driver for the container.

MountPoint
Details on a volume mount point that is used in a container definition.

NetworkBinding
Details on the network bindings between a container and its host container instance. After a task reaches the RUNNING status, manual and automatic host and container port assignments are visible in the networkBindings section of DescribeTasks API responses.

NetworkConfiguration
An object representing the network configuration for a task or service.

NetworkInterface
An object representing the elastic network interface for tasks that use the awsvpc network mode.

PlacementConstraint
An object representing a constraint on task placement. For more information, see Task Placement Constraints in the Amazon Elastic Container Service Developer Guide.

PlacementStrategy
The task placement strategy for a task or service. For more information, see Task Placement Strategies in the Amazon Elastic Container Service Developer Guide.

PortMapping
Port mappings allow containers to access ports on the host container instance to send or receive traffic. Port mappings are specified as part of the container definition.

If using containers in a task with the awsvpc or host network mode, exposed ports should be specified using containerPort. The hostPort can be left blank or it must be the same value as the containerPort.

After a task reaches the RUNNING status, manual and automatic host and container port assignments are visible in the networkBindings section of DescribeTasks API responses.

PutAttributesRequest

PutAttributesResponse

RegisterContainerInstanceRequest

RegisterContainerInstanceResponse

RegisterTaskDefinitionRequest

RegisterTaskDefinitionResponse

RepositoryCredentials
The repository credentials for private registry authentication.

Resource
Describes the resources available for a container instance.

RunTaskRequest

RunTaskResponse

Service
Details on a service within a cluster

ServiceEvent
Details on an event associated with a service.

ServiceRegistry
Details of the service registry.

StartTaskRequest

StartTaskResponse

StopTaskRequest

StopTaskResponse

SubmitContainerStateChangeRequest

SubmitContainerStateChangeResponse

SubmitTaskStateChangeRequest

SubmitTaskStateChangeResponse

Task
Details on a task in a cluster.

TaskDefinition
Details of a task definition.

TaskDefinitionPlacementConstraint
An object representing a constraint on task placement in the task definition.

If you are using the Fargate launch type, task placement constraints are not supported.

For more information, see Task Placement Constraints in the Amazon Elastic Container Service Developer Guide.

TaskOverride
The overrides associated with a task.

Tmpfs
The container path, mount options, and size of the tmpfs mount.

Ulimit
The ulimit settings to pass to the container.

UpdateContainerAgentRequest

UpdateContainerAgentResponse

UpdateContainerInstancesStateRequest

UpdateContainerInstancesStateResponse

UpdateServiceRequest

UpdateServiceResponse

VersionInfo
The Docker and Amazon ECS container agent version information about a container instance.

Volume
A data volume used in a task definition.

VolumeFrom
Details on a data volume from another container in the same task definition.

Enums

CreateClusterError
Errors returned by CreateCluster

CreateServiceError
Errors returned by CreateService

DeleteAttributesError
Errors returned by DeleteAttributes

DeleteClusterError
Errors returned by DeleteCluster

DeleteServiceError
Errors returned by DeleteService

DeregisterContainerInstanceError
Errors returned by DeregisterContainerInstance

DeregisterTaskDefinitionError
Errors returned by DeregisterTaskDefinition

DescribeClustersError
Errors returned by DescribeClusters

DescribeContainerInstancesError
Errors returned by DescribeContainerInstances

DescribeServicesError
Errors returned by DescribeServices

DescribeTaskDefinitionError
Errors returned by DescribeTaskDefinition

DescribeTasksError
Errors returned by DescribeTasks

DiscoverPollEndpointError
Errors returned by DiscoverPollEndpoint

ListAttributesError
Errors returned by ListAttributes

ListClustersError
Errors returned by ListClusters

ListContainerInstancesError
Errors returned by ListContainerInstances

ListServicesError
Errors returned by ListServices

ListTaskDefinitionFamiliesError
Errors returned by ListTaskDefinitionFamilies

ListTaskDefinitionsError
Errors returned by ListTaskDefinitions

ListTasksError
Errors returned by ListTasks

PutAttributesError
Errors returned by PutAttributes

RegisterContainerInstanceError
Errors returned by RegisterContainerInstance

RegisterTaskDefinitionError
Errors returned by RegisterTaskDefinition

RunTaskError
Errors returned by RunTask

StartTaskError
Errors returned by StartTask

StopTaskError
Errors returned by StopTask

SubmitContainerStateChangeError
Errors returned by SubmitContainerStateChange

SubmitTaskStateChangeError
Errors returned by SubmitTaskStateChange

UpdateContainerAgentError
Errors returned by UpdateContainerAgent

UpdateContainerInstancesStateError
Errors returned by UpdateContainerInstancesState

UpdateServiceError
Errors returned by UpdateService

Traits

Ecs
Trait representing the capabilities of the Amazon ECS API. Amazon ECS clients implement this trait.