Struct rusoto_ecs::ContainerInstance

pub struct ContainerInstance {
    pub agent_connected: Option<bool>,
    pub agent_update_status: Option<String>,
    pub attachments: Option<Vec<Attachment>>,
    pub attributes: Option<Vec<Attribute>>,
    pub container_instance_arn: Option<String>,
    pub ec_2_instance_id: Option<String>,
    pub pending_tasks_count: Option<i64>,
    pub registered_at: Option<f64>,
    pub registered_resources: Option<Vec<Resource>>,
    pub remaining_resources: Option<Vec<Resource>>,
    pub running_tasks_count: Option<i64>,
    pub status: Option<String>,
    pub version: Option<i64>,
    pub version_info: Option<VersionInfo>,
}

An EC2 instance that is running the Amazon ECS agent and has been registered with a cluster.

Fields
agent_connected: Option<bool>

This parameter returns true if the agent is connected to Amazon ECS. Registered instances with an agent that may be unhealthy or stopped return false. Only instances connected to an agent can accept placement requests.

agent_update_status: Option<String>

The status of the most recent agent update. If an update has never been requested, this value is NULL.

attachments: Option<Vec<Attachment>>

The elastic network interfaces associated with the container instance.

attributes: Option<Vec<Attribute>>

The attributes set for the container instance, either by the Amazon ECS container agent at instance registration or manually with the PutAttributes operation.

container_instance_arn: Option<String>

The Amazon Resource Name (ARN) of the container instance. The ARN contains the arn:aws:ecs namespace, followed by the Region of the container instance, the AWS account ID of the container instance owner, the container-instance namespace, and then the container instance ID. For example, arn:aws:ecs:region:aws_account_id:container-instance/container_instance_ID .

ec_2_instance_id: Option<String>

The EC2 instance ID of the container instance.

pending_tasks_count: Option<i64>

The number of tasks on the container instance that are in the PENDING status.

registered_at: Option<f64>

The Unix time stamp for when the container instance was registered.

registered_resources: Option<Vec<Resource>>

For CPU and memory resource types, this parameter describes the amount of each resource that was available on the container instance when the container agent registered it with Amazon ECS; this value represents the total amount of CPU and memory that can be allocated on this container instance to tasks. For port resource types, this parameter describes the ports that were reserved by the Amazon ECS container agent when it registered the container instance with Amazon ECS.

remaining_resources: Option<Vec<Resource>>

For CPU and memory resource types, this parameter describes the remaining CPU and memory that has not already been allocated to tasks and is therefore available for new tasks. For port resource types, this parameter describes the ports that were reserved by the Amazon ECS container agent (at instance registration time) and any task containers that have reserved port mappings on the host (with the host or bridge network mode). Any port that is not specified here is available for new tasks.

running_tasks_count: Option<i64>

The number of tasks on the container instance that are in the RUNNING status.

status: Option<String>

The status of the container instance. The valid values are ACTIVE, INACTIVE, or DRAINING. ACTIVE indicates that the container instance can accept tasks. DRAINING indicates that new tasks are not placed on the container instance and any service tasks running on the container instance are removed if possible. For more information, see Container Instance Draining in the Amazon Elastic Container Service Developer Guide.

version: Option<i64>

The version counter for the container instance. Every time a container instance experiences a change that triggers a CloudWatch event, the version counter is incremented. If you are replicating your Amazon ECS container instance state with CloudWatch Events, you can compare the version of a container instance reported by the Amazon ECS APIs with the version reported in CloudWatch Events for the container instance (inside the detail object) to verify that the version in your event stream is current.

version_info: Option<VersionInfo>

The version information for the Amazon ECS container agent and Docker daemon running on the container instance.