Struct rusoto_ecs::Service

pub struct Service {
    pub cluster_arn: Option<String>,
    pub created_at: Option<f64>,
    pub created_by: Option<String>,
    pub deployment_configuration: Option<DeploymentConfiguration>,
    pub deployment_controller: Option<DeploymentController>,
    pub deployments: Option<Vec<Deployment>>,
    pub desired_count: Option<i64>,
    pub enable_ecs_managed_tags: Option<bool>,
    pub events: Option<Vec<ServiceEvent>>,
    pub health_check_grace_period_seconds: Option<i64>,
    pub launch_type: Option<String>,
    pub load_balancers: Option<Vec<LoadBalancer>>,
    pub network_configuration: Option<NetworkConfiguration>,
    pub pending_count: Option<i64>,
    pub placement_constraints: Option<Vec<PlacementConstraint>>,
    pub placement_strategy: Option<Vec<PlacementStrategy>>,
    pub platform_version: Option<String>,
    pub propagate_tags: Option<String>,
    pub role_arn: Option<String>,
    pub running_count: Option<i64>,
    pub scheduling_strategy: Option<String>,
    pub service_arn: Option<String>,
    pub service_name: Option<String>,
    pub service_registries: Option<Vec<ServiceRegistry>>,
    pub status: Option<String>,
    pub tags: Option<Vec<Tag>>,
    pub task_definition: Option<String>,
    pub task_sets: Option<Vec<TaskSet>>,
}

Details on a service within a cluster

Fields
cluster_arn: Option<String>

The Amazon Resource Name (ARN) of the cluster that hosts the service.

created_at: Option<f64>

The Unix timestamp for when the service was created.

created_by: Option<String>

The principal that created the service.

deployment_configuration: Option<DeploymentConfiguration>

Optional deployment parameters that control how many tasks run during the deployment and the ordering of stopping and starting tasks.

deployment_controller: Option<DeploymentController>

The deployment controller type the service is using.

deployments: Option<Vec<Deployment>>

The current state of deployments for the service.

desired_count: Option<i64>

The desired number of instantiations of the task definition to keep running on the service. This value is specified when the service is created with CreateService, and it can be modified with UpdateService.

enable_ecs_managed_tags: Option<bool>

Specifies whether to enable Amazon ECS managed tags for the tasks in the service. For more information, see Tagging Your Amazon ECS Resources in the Amazon Elastic Container Service Developer Guide.

events: Option<Vec<ServiceEvent>>

The event stream for your service. A maximum of 100 of the latest events are displayed.

health_check_grace_period_seconds: Option<i64>

The period of time, in seconds, that the Amazon ECS service scheduler ignores unhealthy Elastic Load Balancing target health checks after a task has first started.

launch_type: Option<String>

The launch type on which your service is running. If no value is specified, it will default to EC2. Valid values include EC2 and FARGATE. For more information, see Amazon ECS Launch Types in the Amazon Elastic Container Service Developer Guide.

load_balancers: Option<Vec<LoadBalancer>>

A list of Elastic Load Balancing load balancer objects, containing the load balancer name, the container name (as it appears in a container definition), and the container port to access from the load balancer.

Services with tasks that use the awsvpc network mode (for example, those with the Fargate launch type) only support Application Load Balancers and Network Load Balancers. Classic Load Balancers are not supported. Also, when you create any target groups for these services, you must choose ip as the target type, not instance. Tasks that use the awsvpc network mode are associated with an elastic network interface, not an Amazon EC2 instance.

network_configuration: Option<NetworkConfiguration>

The VPC subnet and security group configuration for tasks that receive their own elastic network interface by using the awsvpc networking mode.

pending_count: Option<i64>

The number of tasks in the cluster that are in the PENDING state.

placement_constraints: Option<Vec<PlacementConstraint>>

The placement constraints for the tasks in the service.

placement_strategy: Option<Vec<PlacementStrategy>>

The placement strategy that determines how tasks for the service are placed.

platform_version: Option<String>

The platform version on which to run your service. A platform version is only specified for tasks using the Fargate launch type. If one is not specified, the LATEST platform version is used by default. For more information, see AWS Fargate Platform Versions in the Amazon Elastic Container Service Developer Guide.

propagate_tags: Option<String>

Specifies whether to propagate the tags from the task definition or the service to the task. If no value is specified, the tags are not propagated.

role_arn: Option<String>

The ARN of the IAM role associated with the service that allows the Amazon ECS container agent to register container instances with an Elastic Load Balancing load balancer.

running_count: Option<i64>

The number of tasks in the cluster that are in the RUNNING state.

scheduling_strategy: Option<String>

The scheduling strategy to use for the service. For more information, see Services.

There are two service scheduler strategies available:

REPLICA-The replica scheduling strategy places and maintains the desired number of tasks across your cluster. By default, the service scheduler spreads tasks across Availability Zones. You can use task placement strategies and constraints to customize task placement decisions.

DAEMON-The daemon scheduling strategy deploys exactly one task on each container instance in your cluster. When you are using this strategy, do not specify a desired number of tasks or any task placement strategies.

Fargate tasks do not support the DAEMON scheduling strategy.

service_arn: Option<String>

The ARN that identifies the service. The ARN contains the arn:aws:ecs namespace, followed by the Region of the service, the AWS account ID of the service owner, the service namespace, and then the service name. For example, arn:aws:ecs:region:012345678910:service/my-service.

service_name: Option<String>

The name of your service. Up to 255 letters (uppercase and lowercase), numbers, and hyphens are allowed. Service names must be unique within a cluster, but you can have similarly named services in multiple clusters within a Region or across multiple Regions.

service_registries: Option<Vec<ServiceRegistry>>

The details of the service discovery registries to assign to this service. For more information, see Service Discovery.

status: Option<String>

The status of the service. The valid values are ACTIVE, DRAINING, or INACTIVE.

tags: Option<Vec<Tag>>

The metadata that you apply to the service to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

task_definition: Option<String>

The task definition to use for tasks in the service. This value is specified when the service is created with CreateService, and it can be modified with UpdateService.

task_sets: Option<Vec<TaskSet>>

Information about a set of Amazon ECS tasks in either an AWS CodeDeploy or an EXTERNAL deployment. An Amazon ECS task set includes details such as the desired number of tasks, how many tasks are running, and whether the task set serves production traffic.